Swarm

+++ migrations/versions/048_EcrfDetails_Add_redcap_project_id.py
@@ -15,7 +15,8 @@ def upgrade(migrate_engine):
     rp = Table("redcap_project", meta, autoload=True)
     t = Table("ecrf_detail", meta, autoload=True)
 
-    t.c.project_id.drop()
+    if "project_id" in t.c:
+        t.c.project_id.drop()
 
     redcap_project_id = Column("redcap_project_id", Integer, index=True)
     redcap_project_id.create(t, index_name='idx__ecrf_detail__redcap_project_id')
